split particular services
