+++ rules/privatization/config/config.php
@@ -13,5 +13,5 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         ->autoconfigure();
 
     $services->load('Rector\Privatization\\', __DIR__ . '/../src')
-        ->exclude([__DIR__ . '/../src/Rector']);
+        ->exclude([__DIR__ . '/../src/Rector', __DIR__ . '/../src/ValueObject']);
 };
