+++ rules/privatization/src/ValueObject/ConstantVisibility.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\SOLID\ValueObject;
+namespace Rector\Privatization\ValueObject;
 
 final class ConstantVisibility
 {
